# fix: replace node buffers with uint8arrays