--- a/test/diagnose.spec.js
+++ b/test/diagnose.spec.js
@@ -1,9 +1,7 @@
 /* eslint-env mocha */
 'use strict'
 
-const chai = require('chai')
-chai.use(require('dirty-chai'))
-const expect = chai.expect
+const { expect } = require('aegir/utils/chai')
 
 const cbor = require('../')
 const cases = require('./fixtures/cases')