--- a/test/garbage.spec.js
+++ b/test/garbage.spec.js
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 'use strict'
 
-const expect = require('chai').expect
+const { expect } = require('aegir/utils/chai')
 const garbage = require('garbage')
 
 const cbor = require('../src')