All use of node Buffers have been replaced with Uint8Arrays

BREAKING CHANGES:

- `cbor.encode` used to return a Buffer, now it returns a Uint8Array